feat: add TokenDetailHeader component to MarketDetailV2
- Introduced TokenDetailHeader component for enhanced token detail display in MarketDetailV2.
- Integrated TokenDetailHeader into MarketDetailV2, replacing the previous header implementation for improved modularity and maintainability.
- Updated component imports and structure to support the new TokenDetailHeader integration.
