+++ packages/kit/src/views/Market/MarketDetailV2/components/TokenDetailHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './TokenDetailHeader';
