+++ packages/kit/src/views/Market/MarketDetailV2/components/index.ts
@@ -1,3 +1,4 @@
 export * from './MarketDetailHeader';
 export * from './SwapPanel';
 export * from './TokenPriceChart';
+export * from './TokenDetailHeader';
